Move tests to be with similar ones
